Deleted retrofit instance, moved my files in specific folders
