+++ src/android_app/CarsData/app/src/main/java/com/poc/p_couds/APIClient.java
@@ -4,6 +4,8 @@ import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 
+import com.poc.p_couds.models.IApiService;
+
 import java.io.IOException;
 
 import okhttp3.Interceptor;
