--- src/android_app/CarsData/app/src/main/java/com/poc/p_couds/RetrofitInstance.kt
@@ -1,16 +1,0 @@
-package com.poc.p_couds
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitInstance {
-    private const val BASE_URL = "http://10.0.2.2:5000/"
-
-    val api: IApiService by lazy {
-        Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(IApiService::class.java)
-    }
-}
